feat: packed completions, XDG paths, and shell-contract fixes
Complete replies are variable-length (up to 256). Socket lives under
XDG_RUNTIME_DIR and state under XDG_CACHE_HOME. Tab inserts the typed
token instead of gluing ~ paths. File commands allow empty-token
complete; sudo/make/python only on path-like tokens. CLI/helper resolve
from PATH then ~/.local/bin. Helper queries time out. archaic-cli doctor
prints socket, ping, and health.

+++ src/trie.h
@@ -19,8 +19,8 @@ typedef struct RadixNode {
     size_t key_len;
     RadixChild inline_storage[RADIX_INLINE_CHILDREN];
     RadixChild* children;
-    uint8_t child_count;
-    uint8_t child_capacity;
+    uint16_t child_count;
+    uint16_t child_capacity;
     uint64_t freq;
     uint64_t last_access;
     bool is_leaf;
